Add generic solution to decide to double buffer
Kartik's existing solution assumed the vector unit processes one-byte
elements. Our vector unit operates on two-byte BFloat16 elements. The
code below generalizes the solution to work for any datatype.

The core logic is the same. We check to see if the input and output
ports on the vector unit would bottleneck us, assuming we blasted one
vector per cycle through it. If it can handle that, we use the direct
path. Otherwise, we use the double buffered path.